+++ tracer/bigsim/TraceReader.h
@@ -27,7 +27,7 @@ public:
     void loadOffsets();
     void loadTraceSummary();
     void readTrace(int* tot, int* numnodes, int* empes, int* nwth, PE* pe,
-        int penum, int jobnum, unsigned long long* startTime);
+        int penum, int jobnum, double* startTime);
     void setTaskFromLog(Task *t, BgTimeLog* bglog, int taskPE, int emPE, int jobPEindex, PE* pe);
 
     int numEmPes;	// number of emulation PEs, there is a trace file for each of them
